Implementing Code Review Feedback from Daniel Silva
